fully removing test-connection
